filling-machine/requirements.txt
minimalmodbus>=2.0.3
paho-mqtt>=1.6.1
pytest>=7.0.0